Add time and scramble editing in solve dialog

+++ app/src/main/java/com/pluscubed/plustimer/ui/HistorySolveListActivity.java
@@ -39,11 +39,9 @@ public class HistorySolveListActivity extends Activity
 
     @Override
     public void onDialogDismissed(String displayName, int sessionIndex, int solveIndex,
-                                  int penalty) {
-        switch (penalty) {
-            case SolveDialogFragment.DIALOG_RESULT_DELETE:
-                PuzzleType.get(displayName).getSession(sessionIndex).deleteSolve(solveIndex);
-                break;
+                                  boolean delete) {
+        if (delete) {
+            PuzzleType.get(displayName).getSession(sessionIndex).deleteSolve(solveIndex);
         }
 
         if (getSolveListFragment() != null) {
